Added page navs in header
The navs fall under the header if the screen is
narrow. Also made page navs a bit clearer to
use in narrow viewports by avoiding collapse of
the titles. Centered the page icon when
there's nothing else on the header bar. And
made a default "/forecast" route that redirects
to the latest federal forecast.